requirements.txt
Django==4.0.5
djangorestframework==3.13.1
drf_spectacular==0.22.1
PyJWT==2.4.0
PyMySQL==1.0.2
